requirements.txt
django==2.0
django-import-export
django-mptt
django-tinymce
django-widget-tweaks == 1.4.1

python-slugify
django-model-utils

Pillow
python-dateutil
psycopg2
django-grappelli
djangorestframework